+++ backend/.env.sample
@@ -2,6 +2,6 @@ APP_PORT=6000
 FRONTEND_URL=http://localhost:3000
 DB_HOST=localhost
 DB_PORT=3306
-DB_USER=notroot
-DB_PASSWORD=helloworld
-DB_NAME=mvc_express
+DB_USER=user
+DB_PASSWORD=password
+DB_NAME=serenity
