+++ frontend/src/App.jsx
@@ -26,7 +26,7 @@ import SecretariatInterventionPage from "./pages/secretariat/SecretariatInterven
 import SecretariatPatientPage from "./pages/secretariat/SecretariatPatientPage";
 import SecretariatPractitionerPage from "./pages/secretariat/SecretariatPractitionerPage";
 import SecretariatUnitPage from "./pages/secretariat/SecretariatUnitPage";
-import AddPatient from "./components/secretariat/addPatient/AddPatient";
+import AddPatient from "./components/secretariat/addPatient/AddPatientForm";
 import AddPractitioner from "./components/secretariat/addPractitioner/AddPractitioner";
 import AddIntervention from "./components/secretariat/addIntervention/AddIntervention";
